+++ src/exception.h
@@ -25,25 +25,29 @@
 #include <map>
 #include <string>
 
-typedef std::map<unsigned long, std::string> FunctionMap;
+typedef std::map<uint32_t, std::string> FunctionMap;
 
-struct SEHChain{
-	SEHChain *prev;
-	void *SEHfunction;
+struct SEHChain {
+    SEHChain *prev;
+    void *SEHfunction;
 };
 
-class ExceptionHandler
-{
+class ExceptionHandler {
 public:
-	ExceptionHandler();
-	~ExceptionHandler();
-	bool InstallHandler();
-	bool RemoveHandler();
+    ExceptionHandler();
+
+    ~ExceptionHandler();
+
+    bool InstallHandler();
+
+    bool RemoveHandler();
+
 private:
-	bool LoadMap();
-	bool installed;
-	SEHChain chain;
-	
+    bool LoadMap();
+
+    bool installed;
+    SEHChain chain;
+
 };
 
 #endif  // #ifndef __EXCEPTION_H__
